+++ src/app/service/analytic.service.ts
@@ -1,7 +1,8 @@
 import {Injectable} from "@angular/core";
 import {HttpClient} from "@angular/common/http";
 import {Observable} from "rxjs";
-import {MainSession, UserAction} from "../model/trace.model";
+import {AnalyticDto} from "../model/request.model";
+import {UserAction} from "../model/trace.model";
 
 @Injectable({ providedIn: 'root' })
 export class AnalyticService {
@@ -10,13 +11,13 @@ export class AnalyticService {
     constructor(private http: HttpClient) {
     }
 
-    getUserActionsByUser(user: string, date: Date, offset: number, limit: number): Observable<MainSession[]> {
+    getUserActionsByUser(user: string, date: Date, offset: number, limit: number): Observable<AnalyticDto[]> {
         let params: any = {
             'date': date.toISOString()
         }
         if(offset) params['offset'] = offset;
         if(limit) params['limit'] = limit;
-        return this.http.get<Array<MainSession>>(`${this.server}/session/user/${user}/action`, { params: params });
+        return this.http.get<AnalyticDto[]>(`${this.server}/session/user/${user}/action`, { params: params });
     }
 
     getUserActionsBySession(idSession: string): Observable<UserAction[]> {
