+++ src/app/service/jquery/ldap-request.service.ts
@@ -1,9 +1,8 @@
-import { Injectable } from "@angular/core";
-import { HttpClient } from "@angular/common/http";
-import {forkJoin, map, Observable} from "rxjs";
-import { LdapMainExceptionsByPeriodAndappname, LdapSessionExceptionsByPeriodAndappname } from "src/app/model/jquery.model";
-import { NamingRequest} from "../../model/trace.model";
-import {DirectoryRequestDto} from "../../model/new/request.model";
+import {Injectable} from "@angular/core";
+import {HttpClient} from "@angular/common/http";
+import {Observable} from "rxjs";
+import {LdapSessionExceptionsByPeriodAndappname} from "src/app/model/jquery.model";
+import {DirectoryRequestDto} from "../../model/request.model";
 
 
 @Injectable({ providedIn: 'root' })
