+++ src/app/service/jquery/rest-request.service.ts
@@ -1,12 +1,8 @@
-import { Injectable } from "@angular/core";
-import { HttpClient } from "@angular/common/http";
-import {forkJoin,map, Observable} from "rxjs";
-import {
-    RestMainExceptionsByPeriodAndappname,
-    RestSessionExceptionsByPeriodAndappname
-} from "src/app/model/jquery.model";
-import {RestRequest} from "../../model/trace.model";
-import {RestRequestDto} from "../../model/new/request.model";
+import {Injectable} from "@angular/core";
+import {HttpClient} from "@angular/common/http";
+import {Observable} from "rxjs";
+import {RestSessionExceptionsByPeriodAndappname} from "src/app/model/jquery.model";
+import {RestRequestDto} from "../../model/request.model";
 
 
 @Injectable({ providedIn: 'root' })
