+++ src/app/views/detail/session/_component/database-table/detail-database-table.component.ts
@@ -4,7 +4,7 @@ import {MatSort} from "@angular/material/sort";
 import {MatTableDataSource} from "@angular/material/table";
 import {DatePipe} from "@angular/common";
 import {INFINITY} from "../../../../constants";
-import {DatabaseRequestDto} from "../../../../../model/new/request.model";
+import {DatabaseRequestDto} from "../../../../../model/request.model";
 
 @Component({
     selector: 'database-table',
