+++ src/app/views/detail/session/_component/local-table/detail-local-table.component.ts
@@ -2,7 +2,7 @@ import {Component, Input, ViewChild} from "@angular/core";
 import {MatTableDataSource} from "@angular/material/table";
 import {MatPaginator} from "@angular/material/paginator";
 import {MatSort} from "@angular/material/sort";
-import {LocalRequest} from "../../../../../model/new/trace.model";
+import {LocalRequest} from "../../../../../model/trace.model";
 
 @Component({
     selector: 'local-table',
